% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_mean.R
\name{calculate_mean}
\alias{calculate_mean}
\title{Calculate Mean of a Numeric Vector}
\usage{
calculate_mean(data)
}
\arguments{
\item{data}{A numeric vector.}
}
\value{
The mean of the input data.
}
\description{
This function calculates the mean of a numeric vector.
}
\examples{
calculate_mean(c(1, 2, 3, 4, 5))
}
