% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_observations.R
\name{count_observations}
\alias{count_observations}
\title{Count Non-Missing Observations}
\usage{
count_observations(data)
}
\arguments{
\item{data}{A numeric vector.}
}
\value{
The number of non-missing values.
}
\description{
This function counts the number of non-missing values in a numeric vector.
}
\examples{
count_observations(c(1, 2, NA, 4, NA))
}
