% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_data.R
\name{plotting_data}
\alias{plotting_data}
\title{Plot Point Data}
\usage{
plotting_data(data, columnA, columnB, colp = "red")
}
\arguments{
\item{data}{A Data frame}

\item{columnA}{Column for the X Axis}

\item{columnB}{Column for the B Axis}

\item{colp}{Color for the points}
}
\value{
A plot
}
\description{
This function plots points for ColumnA and ColumnB with the color ColP
}
\examples{
data = data.frame(A = rnorm(100), B = rnorm(100), C = rnorm(100))

plotting_data(data, 'A', 'B', 'blue')

plotting_data(data, 'B', 'A', 'black')
}
